+++ tests/testthat/test-problems.R
@@ -106,3 +106,20 @@ test_that("problems that are generated more than once are not duplicated", {
   expect_equal(probs$col, 1)
   expect_equal(probs$expected, "an integer")
 })
+
+test_that("problems return the proper row number", {
+  expect_warning(
+    x <- vroom(I("a,b,c\nx,y,z,,"), altrep = FALSE)
+  )
+  expect_equal(problems(x)$row, 2)
+
+  expect_warning(
+    y <- vroom(I("a,b,c\nx,y,z\nx,y,z,,"), altrep = FALSE)
+  )
+  expect_equal(problems(y)$row, 3)
+
+  expect_warning(
+    z <- vroom(I("a,b,c\nx,y,z,\nx,y,z,,"), altrep = FALSE)
+  )
+  expect_equal(problems(z)$row, c(2, 3))
+})